requirements.txt
scrapy_proxies==0.4
